simple unit test
