+++ vite.config.ts
@@ -1,6 +1,11 @@
 import { sveltekit } from '@sveltejs/kit/vite';
-import { defineConfig } from 'vite';
+// same as import from vite/config, but with vitest types
+import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
-	plugins: [sveltekit()]
+	plugins: [sveltekit()],
+	test: {
+    globals: true,
+    environment: 'jsdom',
+  },
 });
